Cleanup.

+++ examples/38-bloom/bloom.cpp
@@ -395,16 +395,30 @@ public:
 						m_texChainFb[ii]  = bgfx::createFrameBuffer(
 							  (uint16_t)(m_width  >> ii)
 							, (uint16_t)(m_height >> ii)
-							, bgfx::TextureFormat::RGBA32F
+							, bgfx::TextureFormat::RGBA16F
 							, tsFlags
 							);
 					}
 
 					bgfx::TextureHandle gbufferTex[] =
 					{
-						bgfx::createTexture2D(uint16_t(m_width), uint16_t(m_height), false, 1, bgfx::TextureFormat::RGBA32F, tsFlags),
+						bgfx::createTexture2D(
+							  uint16_t(m_width)
+							, uint16_t(m_height)
+							, false
+							, 1
+							, bgfx::TextureFormat::RGBA16F
+							, tsFlags
+							),
 						bgfx::getTexture(m_texChainFb[0]),
-						bgfx::createTexture2D(uint16_t(m_width), uint16_t(m_height), false, 1, bgfx::TextureFormat::D32F, tsFlags),
+						bgfx::createTexture2D(
+							  uint16_t(m_width)
+							, uint16_t(m_height)
+							, false
+							, 1
+							, bgfx::TextureFormat::D32F
+							, tsFlags
+							),
 					};
 
 					m_gbuffer = bgfx::createFrameBuffer(BX_COUNTOF(gbufferTex), gbufferTex, true);
